CustomCell_1_2_3_pg23
Leehaeyong

+++ EX_1_2_1/CartItem.h
@@ -13,6 +13,6 @@
 
 @property (weak, nonatomic) Product* product;
 
-@property (nonatomic) NSInteger quantity;
+@property (nonatomic) int quantity;
 
 @end
